+++ checkout-ui-custom/webpack.config.js
@@ -2,10 +2,7 @@ const path = require('path');
 
 module.exports = {
   mode: process.env.NODE_ENV === 'production' ? 'production' : 'development',
-  entry: [
-    './src/checkout6-custom.js',
-    './src/checkout6-custom.scss'
-  ],
+  entry: ['./src/checkout6-custom.js', './src/checkout6-custom.scss'],
   output: {
     filename: 'checkout6-custom.js',
     path: path.resolve(__dirname, './')
@@ -29,6 +26,26 @@ module.exports = {
           },
           {
             loader: 'sass-loader'
+          }
+        ]
+      },
+      {
+        test: /\.css$/,
+        use: ['style-loader', 'css-loader']
+      },
+      {
+        test: /\.js$/,
+        loader: 'babel-loader',
+        exclude: /node_modules/
+      },
+      {
+        test: /\.(png|jpg|gif)$/i,
+        use: [
+          {
+            loader: 'url-loader',
+            options: {
+              limit: false
+            }
           }
         ]
       }